+++ src/index.js
@@ -1,3 +1,12 @@
 export Container from './container';
 export Row from './row';
 export Col from './col';
+export {
+  breakpointNext,
+  breakpointMin,
+  breakpointMax,
+  mediaBreakpointUp,
+  mediaBreakpointDown,
+  mediaBreakpointBetween,
+  mediaBreakpointOnly
+} from './breakpoints';