+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [4.0.0]
+
+### Changed
+
+- rename `styles.gridBreakpoints` to `styles.breakpoints`
+- rename `styles.gridColumns` to `styles.columns`
+- rename `styles.gridGutterWidth` to `styles.gutterWidth`
+
+### Added
+
+- export breakpoint helper functions
+
 ## [3.0.0]
 
 ### Changed